Move logic to an auto rehidrating place.
